requirements.txt
bentoml
pandas
scikit-learn
grpcio
grpcio-tools
numpy
pydantic
typing_extensions
pytest
